refactor: united connection definition
